requirements.txt
absl-py==2.0.0
albumentations==0.4.3
blobfile==1.0.5
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.6
einops==0.7.0
filelock==3.13.1
fonttools==4.50.0
fsspec==2023.10.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
grpcio==1.60.0
idna==3.6
imageio==2.33.0
imgaug==0.2.6
importlib-metadata==7.0.0
importlib_resources==6.4.0
Jinja2==3.1.2
joblib==0.14.0
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
piq==0.7.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodomex==3.19.0
pyparsing==3.1.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
six==1.16.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tifffile==2023.9.26
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
Werkzeug==3.0.1
xmltodict==0.12.0
zipp==3.17.0
